\name{climatedata}
\alias{climatedata}
\docType{data}
\title{
Climate data
}
\description{
A dataset containing the hourly temperatures 
}
\usage{data("climatedata")}
\format{
  A data frame with 5003 rows and 3 variables.
  \describe{
   \item{date}{Date, in Year-Month-Day format}
   \item{hora}{Hour in numeric format}
   \item{temp}{Temperature, in Celsius-Degree}
  }
}
\keyword{datasets}
