% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_down.R
\name{count_down}
\alias{count_down}
\title{Count Down}
\usage{
count_down(tl, x, total = TRUE)
}
\arguments{
\item{tl}{upper limit}

\item{x}{Vector containing values (data).}

\item{total}{TRUE Shows the total value of hours.}
}
\value{
The function returns the total value of hours.
}
\description{
Function to quantify temperature hours below x.
}
\details{
Function to quantify temperature hours below x.
}
\examples{

x <- rnorm(1000, 15, 5)

count_down(tl = 7.2, x = x, total = TRUE)

}
