% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_up.R
\name{count_up}
\alias{count_up}
\title{Count Up}
\usage{
count_up(tl, x, total = TRUE)
}
\arguments{
\item{tl}{lower limit}

\item{x}{Vector containing values (data).}

\item{total}{TRUE Shows the total value of hours.}
}
\value{
The function returns the total value of hours.
}
\description{
Function to quantify temperature hours above x.
}
\details{
Function to quantify temperature hours above x.
}
\examples{
count_up(tl = 30, x = rnorm(1000, 20, 10), total = TRUE)

}
\author{
Marcos Robson Sachet & Rafael Henrique Pertille.
}
