% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_model.R
\name{dynamic_model}
\alias{dynamic_model}
\title{Dynamic Model}
\usage{
dynamic_model(x, total = TRUE)
}
\arguments{
\item{x}{Vector containing temperature values (Celsius-degree).}

\item{total}{TRUE Shows the total value of accumulation, FALSE shows the value of chill for each temperature (TRUE is default).}
}
\value{
The function returns values the chill for each temperature of vector (Total = FALSE), or returns the chill accumulation (Total = TRUE).
}
\description{
Quantifies the chill accumulation rates with dynamic equation.
}
\details{
The model is based on dynamic accumulation, by means of the relationship between temperatures. The dynamic model assumes that the accumulated chill-units are not annulled by high temperatures.
}
\references{
FISHMAN, Svetlana, EREZ, A. & COUVILLON, G. A. (1987). The Temperature Dependence of Dormancy Breaking in Plants: Computer Simulation of Processes Studied Under Controlled Temperatures. J. Theor. Biol.
}
\author{
Marcos Robson Sachet & Rafael Henrique Pertille.
}
