% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gala_model.R
\name{gala_model}
\alias{gala_model}
\title{Gala model}
\usage{
gala_model(x, total = TRUE)
}
\arguments{
\item{x}{Vector containing temperature values (Celsius-degree).}

\item{total}{TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).}
}
\value{
The function returns values the chill-units for each temperature of vector (Total = FALSE), or returns the chill-units accumulation (Total = TRUE).
}
\description{
Quantifies the chill accumulation by means of converting temperatures to chill-units.
}
\details{
The model is based on chill-units, where 1 chill-unit is when the tree is exposure between -2.1°C and 5.5°C. When the temperature is above 18°C, the chill-unit is -1. The chill-units accumulation is 0 when occurs temperature between 13°C and 16°C.
}
\note{
This model was make for quantifying the chill accumulation rates to 'Gala' apple or varieties of the Gala group. We aren't recommended the application for others species or others groups of the apple tree.
}
\examples{

x <- rnorm(500, 10, 5)
gala_model(x)
gala_model(x, FALSE)

}
\references{
GUAK, Sunghee & NEILSEN, Denise. (2013). Chill Unit Models for Predicting Dormancy Completion of Floral Buds in Apple and Sweet Cherry.
}
