% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_dynamic.R
\name{modified_dynamic}
\alias{modified_dynamic}
\title{Dynamic Model Modified}
\usage{
modified_dynamic(x, total = TRUE)
}
\arguments{
\item{x}{Vector containing temperature values (Celsius-degree).}

\item{total}{TRUE Shows the total value of accumulation, FALSE shows the value of chill for each temperature (TRUE is default).}
}
\value{
The function returns values the chill for each temperature of vector (Total = FALSE), or returns the chill accumulation (Total = TRUE).
}
\description{
Quantifies the chill accumulation rates with modified dynamic equation for subtropical climate conditions .
}
\details{
The model is based on dynamic accumulation, by means of the relationship between temperatures. It is applicable in subtropical climate conditions. This model was calibrated and validated based on observations of orchards in Palmas - PR, Brazil.
}
\examples{

x <- rnorm(500, 15, 8)
modified_dynamic(x)
modified_dynamic(x, FALSE)

}
\references{
FISHMAN, Svetlana, EREZ, A. & COUVILLON G. A. (1987). The Temperature Dependence of Dormancy Breaking in Plants: Computer Simulation of Processes Studied Under Controlled Temperatures. J. Theor. Biol.
}
\author{
Marcos Robson Sachet, Idemir Citadin & Rafael Henrique Pertille
}
