% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pcu.R
\name{pcu_model}
\alias{pcu_model}
\title{Positive Chill Units - PCU}
\usage{
pcu_model(x, total = TRUE)
}
\arguments{
\item{x}{Vector containing temperature values (Celsius-degree).}

\item{total}{TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).}
}
\value{
The function returns values the chill-units for each temperature of vector.
}
\description{
Quantifies the chill and heat accumulation rates by means of converting temperatures to positive chill-units - Modified Utah Model.
}
\details{
The PCU Model is the modified Utah Model. When the temperature is above 15.9°C, the chill-unit is 0. This modification was made because when the Utah model is applied in warm conditions, accumulation becomes negative.
}
\examples{

a <- pcu_model(rnorm(1000, 15, 4))
hist(a)

}
\references{
Richardson, E, A. et al. 1974. "A Model for Estimating the Completation of Rest for 'Redhaven' and 'Elberta' Peach Trees". Research Reports & Notes. 
EBERT, A. et al. 1986. First experiences with chill units models is southern Brazil. Acta Horticulturae, The Hague,v. 184, p. 89-96.
}
