% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_between.R
\name{count_between}
\alias{count_between}
\title{Count between}
\usage{
count_between(tli, tls, x, total = TRUE)
}
\arguments{
\item{tli}{lower limit}

\item{tls}{upper limit}

\item{x}{Vector containing values (data).}

\item{total}{TRUE Shows the total value of hours.}
}
\value{
The function returns the total value of hours.
}
\description{
Function to quantify temperature hours between x and y.
}
\details{
Function to quantify temperature hours between x and y.
}
\examples{

x <- rnorm(500, 7, 3)

count_between(tli = 8, tls = 15, x = x,  total = TRUE)

}
