% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unified_model.R
\name{unified_model}
\alias{unified_model}
\title{Unified Model}
\usage{
unified_model(x, total = TRUE)
}
\arguments{
\item{x}{Vector containing temperature values (Celsius-degree).}

\item{total}{TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).}
}
\value{
The function returns values the chill-units for each temperature of vector (Total = FALSE), or returns the chill-units accumulation (Total = TRUE).
}
\description{
Quantifies the chill accumulation rates by means of converting temperatures to chill-units. the coefficients used in the model are for the apple tree.
}
\details{
The model is based on chill-units, but the limits are unknown.
}
\examples{

unified_model(rnorm(100, 10, 3), FALSE)

}
\references{
Chuine, I. et al. 2016. Can phenological models predict tree phenology accurately in the future? The unrevealed hurdle of endodormancy break. Global Change Biology.
}
